Update ApiResource.cs
This clone method is being also used in the Extensions class IResourceStoreExtensions in the method
public static async Task<Resources> FindResourcesByScopeAsync(this IResourceStore store, IEnumerable<string> scopeNames)

This call is used in the consent controller. But because the description and display name isn't cloned you're presented with a blanc consent option.
I wasn't able to build it with this fix however.